+++ tests/subworkflows/test_patch.py
@@ -302,6 +302,7 @@ class TestSubworkflowsPatch(TestSubworkflows):
 
         with mock.patch.object(nf_core.components.patch.questionary, "confirm") as mock_questionary:
             mock_questionary.unsafe_ask.return_value = True
+            patch_obj.no_prompts = False
             patch_obj.remove("bam_sort_stats_samtools")
         # Check that the diff file has been removed
         assert not (subworkflow_path / "bam_sort_stats_samtools.diff").exists()